requirements.txt
python-telegram-bot==20.6
yfinance==0.2.31
pandas==2.1.2
numpy==1.26.1
psycopg2-binary==2.9.9
Flask==3.0.0
matplotlib==3.8.1
seaborn==0.13.0
psutil==5.9.6
python-dotenv==1.0.0
